chore: restructure `state.rs`
